feat: refactoring
renaming folders and reduce number of dir levels
