+++ src/read_namespace/main.py
@@ -1,6 +1,6 @@
-from .read_github_namespace.graph.fetch_org_repos import collate, \
+from .query_github.fetch_org_repos import collate, \
     fetch_org_repos
-from .clone.shell_clone import clone_list
+from .shell_cmds.shell_clone import clone_list
 from .arg_parsing.parse_cmd_line import parse_cmd_line
 from tabulate import tabulate
