+++ src/read_namespace/query_github/.graphqlconfig
