+++ src/read_namespace/query_github/login.graphql
